feat(kc_api): イベント海域情報と航空戦スプライト指標をモデルへ追加
変更内容:\n- battle系モデルに schema_v0_5 向けのスプライト派生指標を追加\n  - f/e_sprite_fly_count\n  - f/e_sprite_crash_stage1_count\n  - f/e_sprite_crash_stage2_count\n  - 一部経路は crash_count として集約\n- 航空戦ダメージ配列を float から int に統一し、変換処理を into_i32 へ揃えた\n- cell/map 変換経路にイベント海域ゲージ情報を追加\n  - event_map_max_maphp / now_maphp / dmg\n  - gauge_type / gauge_num / state / selected_rank\n- ship 変換に cond_state を追加し、クライアント表示閾値に沿う判定関数を導入\n- mst_map_info に gauge_type/gauge_num を追加してマスタ表現を拡張\n- shield判定に関するコードコメントを実装挙動に合わせて明確化\n- feature 構成を整理\n  - kc_api workspace で kc-api-database を default-features=false に変更\n  - kc-api crate の default に schema_v0_5 を追加\n  - kc-api crate に full feature(= kc-api-database/full)を追加\n  - integration-test 側は schema_v0_5 を明示有効化\n\n意図:\n- イベント海域の状態変化と演出情報を、DB/DTO/Interfaceで一貫して保持する\n- 解析用データを画面演出の実態に近い粒度で取り扱えるようにする\n- feature 解決の曖昧さを減らし、利用側で必要機能を明示的に選べるようにする

+++ packages/kc_api/crates/kc-api-dto/src/common/common_battle.rs
@@ -18,7 +18,7 @@ pub struct ApiRaigeki {
     #[serde(rename = "api_fdam")]
     pub api_fdam: Vec<f32>,
     #[serde(rename = "api_fydam")]
-    pub api_fydam: Vec<i64>,
+    pub api_fydam: Vec<f32>,
     #[serde(rename = "api_erai")]
     pub api_erai: Vec<i64>,
     #[serde(rename = "api_ecl")]
@@ -26,7 +26,7 @@ pub struct ApiRaigeki {
     #[serde(rename = "api_edam")]
     pub api_edam: Vec<f32>,
     #[serde(rename = "api_eydam")]
-    pub api_eydam: Vec<i64>,
+    pub api_eydam: Vec<f32>,
 }
 
 #[derive(FieldSizeChecker, TraitForTest)]
@@ -86,7 +86,7 @@ pub struct ApiOpeningAtack {
     #[serde(rename = "api_fdam")]
     pub api_fdam: Vec<f32>,
     #[serde(rename = "api_fydam_list_items")]
-    pub api_fydam_list_items: Vec<Option<Vec<i64>>>,
+    pub api_fydam_list_items: Vec<Option<Vec<f32>>>,
     #[serde(rename = "api_erai_list_items")]
     pub api_erai_list_items: Vec<Option<Vec<i64>>>,
     #[serde(rename = "api_ecl_list_items")]
@@ -94,7 +94,7 @@ pub struct ApiOpeningAtack {
     #[serde(rename = "api_edam")]
     pub api_edam: Vec<f32>,
     #[serde(rename = "api_eydam_list_items")]
-    pub api_eydam_list_items: Vec<Option<Vec<i64>>>,
+    pub api_eydam_list_items: Vec<Option<Vec<f32>>>,
 }
 
 #[derive(FieldSizeChecker, TraitForTest)]
